+++ src/main.rs
@@ -1,6 +1,6 @@
 #![no_std]
 #![no_main]
-#![allow(clippy::unusual_byte_groupings)]
+#![allow(clippy::unusual_byte_groupings, clippy::deref_addrof)]
 #![forbid(static_mut_refs)]
 #![feature(abi_x86_interrupt)]
 
@@ -19,6 +19,9 @@ mod speaker;
 
 /// Handles post-boot startup tasks.
 mod startup;
+
+/// Handles system information.
+mod sysinfo;
 
 /// Handles the PIT.
 mod time;
@@ -40,6 +43,8 @@ pub unsafe extern "C" fn kmain() -> ! {
     startup::run("Initialised PIC", interrupts::init_pic);
     startup::run("Set PIT frequency", time::set_timer_interval);
     startup::run("Initialised keyboard", interrupts::init_kbd);
+    startup::run("Prepared for RTC sync", time::setup_rtc_int);
+    startup::run("Checked CPUID", sysinfo::check_cpuid);
 
     startup::SYS_INIT.store(true, core::sync::atomic::Ordering::Relaxed);
     vga::print_color("All startup tasks completed! \u{1}\n\n", vga::Color::Green);