+++ src/time.rs
@@ -1,4 +1,5 @@
 use crate::{
+    interrupts,
     ports::{self, Port},
     startup,
     vga::Corner,
@@ -6,7 +7,9 @@ use crate::{
 use core::{
     arch::asm,
     convert::Infallible,
-    sync::atomic::{AtomicBool, AtomicU64, Ordering},
+    fmt::Display,
+    mem,
+    sync::atomic::{AtomicBool, AtomicU8, AtomicU64, Ordering},
 };
 
 /// How many ticks the kernel has been running for.
@@ -19,6 +22,12 @@ pub static PIT_BASE_FREQ: u64 = 1193180;
 
 /// Set when wait is called due to the crash when rebooting from another OS.
 pub static WAITING: AtomicBool = AtomicBool::new(false);
+
+/// The time the kernel was launched.
+pub static LAUNCH_TIME: Time = unsafe { mem::zeroed() };
+
+/// CMOS register B.
+static CMOS_REG_B: u8 = 0x8B;
 
 /// Sets the timer interval in channel 0 to 10 ms.
 pub fn set_timer_interval() -> Result<(), Infallible> {
@@ -98,5 +107,116 @@ pub fn wait(ticks: u64) {
 
         *char = prev;
         WAITING.store(false, Ordering::Relaxed);
+    }
+}
+
+/// Second-precise time value.
+pub struct Time {
+    year: AtomicU8,
+    month: AtomicU8,
+    day: AtomicU8,
+    hour: AtomicU8,
+    min: AtomicU8,
+    sec: AtomicU8,
+}
+
+impl Time {
+    /// Returns the current time in the RTC.
+    /// [`Reference`](https://wiki.osdev.org/CMOS#Getting_Current_Date_and_Time_from_RTC)
+    fn now() -> Self {
+        unsafe {
+            Time {
+                year: AtomicU8::new(read_cmos_reg(0x9)),
+                month: AtomicU8::new(read_cmos_reg(0x8)),
+                day: AtomicU8::new(read_cmos_reg(0x7)),
+                hour: AtomicU8::new(read_cmos_reg(0x4)),
+                min: AtomicU8::new(read_cmos_reg(0x2)),
+                sec: AtomicU8::new(read_cmos_reg(0x0)),
+            }
+        }
+    }
+}
+
+impl Display for Time {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(
+            f,
+            " {}:{}:{} {}/{}/{}",
+            self.hour.load(Ordering::Relaxed),
+            self.min.load(Ordering::Relaxed),
+            self.sec.load(Ordering::Relaxed),
+            self.day.load(Ordering::Relaxed),
+            self.month.load(Ordering::Relaxed),
+            self.year.load(Ordering::Relaxed),
+        )
+    }
+}
+
+/// Returns the current value of CMOS register `reg`.
+/// # Safety
+/// Reads and writes to I/O ports.
+unsafe fn read_cmos_reg(reg: u8) -> u8 {
+    unsafe {
+        ports::writeb(Port::CMOSSelector, reg);
+        ports::readb(Port::CMOSRegister)
+    }
+}
+
+/// Sets up RTC interrupts in IRQ 8.
+pub fn setup_rtc_int() -> Result<(), Infallible> {
+    interrupts::cli();
+
+    // Set bit 6 in register B.
+    unsafe {
+        let prev = read_cmos_reg(CMOS_REG_B);
+        ports::writeb(Port::CMOSSelector, CMOS_REG_B);
+        ports::writeb(Port::CMOSRegister, prev | 0b1000000);
+    }
+
+    interrupts::sti();
+    Ok(())
+}
+
+/// Ran by RTC handler when the update ended interrupt occurs.
+/// [`Reference`](https://wiki.osdev.org/CMOS#The_Real-Time_Clock)
+#[unsafe(no_mangle)]
+extern "C" fn sync_time_to_rtc() {
+    /// The 24 hour / AM PM flag in the hours value.
+    static TWENTY_FOUR_HR_FLAG: u8 = 0b10000000;
+
+    let time = Time::now();
+    let reg_b = unsafe { read_cmos_reg(CMOS_REG_B) };
+    let mut hour = time.hour.load(Ordering::Relaxed);
+
+    // If BCD mode (bit 2 clear), convert values to binary using the formula
+    // Binary = ((BCD / 16) * 10) + (BCD & 0xF)
+    if reg_b != reg_b | 0b100 {
+        bcd_to_bin(&time.sec);
+        bcd_to_bin(&time.min);
+        bcd_to_bin(&time.day);
+        bcd_to_bin(&time.month);
+        bcd_to_bin(&time.year);
+
+        // Preserve 24 hour flag
+        hour = ((hour & 0x0F) + (((hour & 0x70) / 16) * 10)) | (hour & TWENTY_FOUR_HR_FLAG);
+    }
+
+    // If 12 hour time (bit 1 clear and flag set)
+    if (reg_b != reg_b | 0b10) && (hour == hour & TWENTY_FOUR_HR_FLAG) {
+        let hour = ((hour & 0x7F) + 12) % 24;
+        time.hour.store(hour, Ordering::Relaxed);
+    }
+
+    // Store time
+    let relaxed = Ordering::Relaxed;
+    LAUNCH_TIME.year.store(time.year.load(relaxed), relaxed);
+    LAUNCH_TIME.month.store(time.month.load(relaxed), relaxed);
+    LAUNCH_TIME.day.store(time.day.load(relaxed), relaxed);
+    LAUNCH_TIME.min.store(time.min.load(relaxed), relaxed);
+    LAUNCH_TIME.sec.store(time.sec.load(relaxed), relaxed);
+
+    fn bcd_to_bin(val: &AtomicU8) {
+        let bcd = val.load(Ordering::Relaxed);
+        val.store(((bcd / 16) * 10) + (bcd & 0xF), Ordering::Relaxed)
     }
 }